Build fix on ARM.
svn path=/trunk/KDE/kdelibs/; revision=1092055

+++ animations/pixmaptransition.cpp
@@ -137,7 +137,7 @@ void PixmapTransition::updateCurrentTime(int currentTime)
         if (!m_startPixmap.isNull() && !m_targetPixmap.isNull()) {
             m_currentPixmap = Plasma::PaintUtils::transition(m_startPixmap, m_targetPixmap, delta);
         } else if (m_startPixmap.isNull()) {
-            if (qFuzzyCompare(delta, 1.0)) {
+            if (qFuzzyCompare(delta, qreal(1.0))) {
                 m_currentPixmap = m_targetPixmap;
                 return;
             }
